requirements.txt
mediapipe>=0.10.0
openai-whisper>=20231117
google-genai>=0.8.0
python-dotenv>=1.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
ruff>=0.4.0
pytest>=8.0.0
